Move the Venus E Mini definition into its own device file
The Venus E Mini shares no protocol with the other Venus models - its cd=1
response has almost no field names in common - so keeping its definition
inside venus.ts put two unrelated devices in one 3200-line file. It now
lives in src/device/venusEMini.ts, alongside the other per-device files,
and venus.ts is back to covering only HMG/VNSE3/VNSA/VNSD.

Pure code movement: the field mappings, enum tables, raw-field bucket and
schedule-slot handling are unchanged. Verified by diffing the generated
Home Assistant discovery payloads for a VNSEMINI-0 before and after - all
93 entities are byte-for-byte identical.

The data model stays in types.ts, where every other device's types live.

One comment was rewritten rather than moved: it referred to the Venus
family "above", which is no longer in this file.
